operationsOverID.go: - rename addIntToID to getIthFingerID and make it calculate (id + 2^next) mod 2^numOfBitsInID
daemons.go:
- fix StabilizeDaemon(): a node should not notify itself
- minor refactors and adaptations to changes in operationsOverID.go

chordTypes.go:
- change NumOfBitsInID from int32 to int

chord.go:
- improve ChordNode.String()

main.go:
- adapt -m to be int

+++ main.go
@@ -48,7 +48,7 @@ func main() {
 	}
 
 	chordNode, err := chord.NewChordNode(listener, chord.ChordConfig{
-		NumOfBitsInID: int32(*m),
+		NumOfBitsInID: *m,
 	})
 	if err != nil {
 		log.Fatal("creating new Chord node failed")
